+++ ydb/public/sdk/cpp/src/client/impl/internal/retry/retry_sync.h
@@ -183,22 +183,14 @@ protected:
     }
 
     TStatusType RunOperation() override {
-<<<<<<< HEAD
-        TInRetryOperationContextClientGuard guard(this->Client_);
-        if constexpr (TFunctionArgs<TOperation>::Length == 1) {
-            return Operation_(this->Session_.value());
-        } else {
-            return Operation_(this->Session_.value(), this->GetRemainingTimeout());
-        }
-=======
         return InvokeWithOptionalCatch<TStatusType>(this->Settings_.CatchYdbExceptions_, [&]() -> TStatusType {
+            TInRetryOperationContextClientGuard guard(this->Client_);
             if constexpr (TFunctionArgs<TOperation>::Length == 1) {
                 return Operation_(this->Session_.value());
             } else {
                 return Operation_(this->Session_.value(), this->GetRemainingTimeout());
             }
         });
->>>>>>> a37d7861181 (refactored the retry wrappers and remove yql script range handler.)
     }
 
     void Reset() override {